remove unnecessary method, rename method name
