+++ rxwebview/src/main/java/com/github/satoshun/reactivex/webview/RxWebViewClient.java
@@ -8,45 +8,12 @@ import android.webkit.WebViewClient;
 import com.github.satoshun.reactivex.webview.data.RxWebViewClientData;
 import com.github.satoshun.reactivex.webview.internal.ObjectHelper;
 
-import io.reactivex.Completable;
 import io.reactivex.Observable;
 
 /**
  * static factory methods for creating {@linkplain io.reactivex.Completable} and {@linkplain io.reactivex.Observable}
  */
 public class RxWebViewClient {
-
-  /**
-   * @deprecated please use events method
-   * <p>
-   * Create an completable which emits on {@code WebView} WebViewClient#onPageStarted event.
-   * <p>
-   * <em>Warning:</em> The method don't combine other method. if you want to subscribe a onPageStarted and
-   * onPageFinished and other events, please use {@link #events(WebView, WebViewClient)} method.
-   * </p>
-   */
-  @CheckResult @NonNull @Deprecated
-  public static Completable onPageStarted(WebView webView, WebViewClient client) {
-    ObjectHelper.requireNonNull(webView, "webView is null");
-    ObjectHelper.requireNonNull(client, "client is null");
-    return Completable.create(new PageStartedOnSubscribe(webView, client));
-  }
-
-  /**
-   * @deprecated please use events method
-   * <p>
-   * Create an completable which emits on {@code WebView} WebViewClient#onPageFinished event.
-   * <p>
-   * <em>Warning:</em> The method don't combine other method. if you want to subscribe a onPageStarted and
-   * onPageFinished and other events, please use {@link #events(WebView, WebViewClient)} method.
-   * </p>
-   */
-  @CheckResult @NonNull @Deprecated
-  public static Completable onPageFinished(WebView webView, WebViewClient client) {
-    ObjectHelper.requireNonNull(webView, "webView is null");
-    ObjectHelper.requireNonNull(client, "client is null");
-    return Completable.create(new PageFinishedOnSubscribe(webView, client));
-  }
 
   /**
    * @deprecated please use events method
@@ -69,6 +36,6 @@ public class RxWebViewClient {
   public static Observable<RxWebViewClientData> events(WebView webView, WebViewClient client) {
     ObjectHelper.requireNonNull(webView, "webView is null");
     ObjectHelper.requireNonNull(client, "client is null");
-    return Observable.create(new AllOnSubscribe(webView, client));
+    return new AllOnObservable(webView, client);
   }
 }
